melt parser is now a generator
