+++ .gitignore
@@ -143,3 +143,6 @@ dmypy.json
 .venv/
 
 # End of https://www.gitignore.io/api/python
+
+
+*.bb
