--- transposcope/test.py
@@ -1,5 +1,0 @@
-import sys
-import pprint
-import transposcope
-
-pprint.pprint(sys.path)
